requirements.txt
ultralytics
streamlit
opencv-python-headless
python-dotenv